Revert "tips segun status"
