+++ src/index.css
@@ -7,17 +7,15 @@
     box-sizing: border-box;
     margin: 0;
     padding: 0;
-
+    font-family: Poppins, sans-serif;
 }
 
 :root {
-    --font-sans-main: 'Poppins', sans-serif;
-    --font-sans-secondary: 'Noto Sans', sans-serif;
-    --col-white: rgb(246 246, 246);
-    --col-dark: rgb(53 44 53);
-    --col-green-1:rgb(186 235 73) ;
+    --col-white: rgb(246, 246, 246);
+    --col-darkgrey: rgb(53, 44, 53);
+    --col-green-1:rgb(186, 235, 73) ;
     --col-green-2: rgb(145 217 73);
-    --col-yellow-1: rgb(255 228 85);
+    --col-yellow-1: rgb(255, 228, 85);
     --col-yellow-2: rgb(247 217 60);
     --col-orange-1: rgb(253 154 101);
     --col-orange-2: rgb(238 132 75);
@@ -27,4 +25,10 @@
     --col-red-2: rgb(118 38 43);
     --col-violet-1: rgb(201 94 204);
     --col-violet-2: rgb(192 59 195);
+    --col-bg-good: ;
+    --col-bg-medium: ;
+    --col-bg-high:var(--col-bg-bad) ;
+    --col-bg-bad: ;
+    --col-bg-danger: ;
+    --col-bg-death: ;
 }
