+++ src/pages/Home.jsx
@@ -1,18 +1,12 @@
 
-import SearchCity from "../components/searchCity/SearchCity";
-
+import AirCityInfo from "./airQuality/airCityInfo/AirCityInfo";
 
 
 const Home  = () => {
 
     return (
-        <div className="w-screen h-screen bg-white-gradient absolute top-0 z-[-2]">
-            <div className="mt-[11rem]">
-                <h1 className="text-[5rem] font-[600] text-center">SWEET</h1>
-                <h2 className="text-[5rem] font-[600] text-center mt-[-3rem]">Air</h2>
-            </div>
-            <SearchCity className="pt-[5rem] pl-[4.7rem] lg:ml-[36%]" text="Busca tu ciudad" length="w-[15rem]"/>
-            <img src="public/assets/wavesIntroDesktop.png" className="absolute bottom-0 z-[-1] object-right w-[100%]"></img>
+        <div className="w-screen h-screen bg-green-gradient">
+            < AirCityInfo status="buena" qualityNumber="47" city="Oviedo"/>
         </div>
 
     )
