+++ src/pages/Search.jsx
@@ -1,18 +1,7 @@
-import SearchCity from "../components/searchCity/SearchCity";
-import Navbar from "../components/logo/Navbar";
-
-
 const Search = () => {
 
     return (
         <>
-        <div className="w-screen h-screen bg-white-gradient absolute top-0 z-[-2]">
-            <Navbar/>
-            <div className="flex flex-col lg:ml-[36%]">
-            <SearchCity className="pt-[5rem] pl-[4.7rem]" text="Busca tu ciudad" length="w-[15rem]"/>
-            </div>
-            <img src="public/assets/wavesIntroDesktop.png" className="absolute bottom-0 z-[-1] object-right w-[100%]"></img>
-        </div>
         </>
     )
 }
