+++ src/pages/airQuality/AirQualityMain.jsx
@@ -1,0 +1,10 @@
+const AirQualityMain = () => {
+
+    return (
+        <div className="w-screen h-screen bg-gradient-to-t from-lime-400 to-lime-300">
+            
+        </div>
+    )
+}
+
+export default AirQualityMain;
