+++ src/routes/index.jsx
@@ -1,11 +1,10 @@
 import { createBrowserRouter } from "react-router-dom";
 import Layout from "../layout/Layout";
 import Home from "../pages/Home";
-import AirQualityMain from "../pages/AirQualityMain";
+import AirQualityMain from "../pages/airQuality/AirQualityMain";
 import Comments from "../pages/Comments";
 import Login from "../pages/Login";
 import AdminComments from "../pages/AdminComments";
-import Search from "../pages/Search";
 
 
 
@@ -18,7 +17,7 @@ const router = createBrowserRouter ([
                 element: <Home />
             },
             { 
-                path:"/airquality/:cityName",
+                path:"/airquality",
                 element: <AirQualityMain />
             },
             { 
@@ -32,10 +31,6 @@ const router = createBrowserRouter ([
             { 
                 path:"/admin-comments",
                 element: <AdminComments />
-            },
-            { 
-                path:"/search",
-                element: <Search />
             },
         ]
     }
