--- src/utils/tipsData.js
@@ -1,52 +1,0 @@
-const tipsData = [
-    {
-        id: "1",
-        text: "Abraza árboles"
-    },
-    {
-        id: "1",
-        text: "Sal a jugar"
-    },
-    {
-        id: "1",
-        text: "No se anticipan impactos a la salud cuando la calidad del aire se encuentra en este intervalo."
-    },
-    {
-        id: "2",
-        text: "Usa mascarilla"
-    },
-    {
-        id: "2",
-        text: "Las personas extraordinariamente sensitivas deben considerar limitación de los esfuerzos físicos excesivos y prolongados al aire libre."
-    },
-    {
-        id: "3",
-        text: "Usa mascarilla y protege nariz y ojos"
-    },
-    {
-        id: "3",
-        text: "Los niños y adultos activos, y personas con enfermedades respiratorias tales como el asma, deben evitar los esfuerzos físicos excesivos y prolongados al aire libre."
-    },
-    {
-        id: "4",
-        text: "Los niños y adultos activos, y personas con enfermedades respiratorias tales como el asma, deben evitar los esfuerzos excesivos prolongados al aire libre; las demás personas, especialmente los niños, deben limitar los esfuerzos físicos excesivos y prolongados al aire libre."
-    },
-    {
-        id: "4",
-        text: "Usa FPP2 y cúbrete los ojos"
-    },
-    {
-        id: "5",
-        text: "Evita salir de casa"
-    },
-    {
-        id: "5",
-        text: "Los niños y adultos activos, y personas con enfermedades respiratorias tales como el asma, deben evitar todos los esfuerzos excesivos al aire libre; las demás personas, especialmente los niños, deben limitar los esfuerzos físicos excesivos al aire libre."
-    },
-    {
-        id: "6",
-        text: "NO salgas de casa"
-    }
-]
-
-export default tipsData
